+++ docs/changelog.rst
@@ -2,6 +2,18 @@
 Changelog
 =========
 
+
+.. topic::  Version 0.5.6
+
+    * Add dimension_type and sort_order columns to the dimension_description
+    * Add end dates to the observation_fact export in skinny loader (end date dimension still missing)
+
+.. topic::  Version 0.5.5
+
+    * 16.2 template validation
+    * Allow reuse of column names in separate source files used by template reader
+    * Template reader source data can be provided in single sheet Excel files
+
 .. topic::  Version 0.5.4
 
     * Create transmart-copy files without setting ``FAS`` on study node.